+++ src/OpenMonday.Core/Expose/Int/IBoardServices.cs
@@ -1,8 +1,11 @@
 public interface IBoardServices
 {
-    public Task<ServiceResult<T>> RetrieveAndBuildBoard<T, TItem>(string board_id, BoardMapping boardMapping) where T : Board, new() where TItem : Board_Item, new();
+    Task<ServiceResult<T>> RetrieveAndBuildBoard<T, TItem>(string board_id, BoardMapping boardMapping) where T : Board, new() where TItem : Board_Item, new();
 
-    public Task<ServiceResult<T>> RetrieveAndBuildBoard<T, TItem>(string board_id) where T : Board, new() where TItem : Board_Item, new();
+    Task<ServiceResult<T>> RetrieveAndBuildBoard<T, TItem>(string board_id) where T : Board, new() where TItem : Board_Item, new();
 
-    public Task<ServiceResult<MondayDriverBoardStructure>> GetBoardsStructureById(string boardId);
+    Task<ServiceResult<BoardStructure>> RetrieveBoardStructure(string boardId);
+
+    [Obsolete("This method is deprecated, use RetrieveBoardStructure")]
+    Task<ServiceResult<MondayDriverBoardStructure>> GetBoardsStructureById(string boardId);
 }